pyproject.toml
[project]
name = "lcv-retail-analytics-platform"
version = "0.1.0"
description = "Retail analytics platform for LCV data processing, modeling, and insights"
authors = [{name = "Kutsi Tusuz", email = "ktusuz@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "psycopg2-binary==2.9.10",
    "pandas==2.2.3",
    "pyarrow==23.0.0",          # Parquet format support (ETL)
    "google-cloud-storage==2.18.2",
    "google-cloud-bigquery==3.41.0",
    "google-auth==2.49.0",      # GCP service account authentication
    "dbt-core==1.11.8",
    "dbt-bigquery==1.11.1",
    "fastapi==0.115.12",
    "uvicorn==0.32.1",
    "sqlalchemy==2.0.36",
    "scikit-learn==1.5.2",
    "xgboost==2.1.3",
    "python-dotenv==1.0.1",
    "requests==2.33.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black==25.12.0",
    "ruff==0.15.1",
    "mypy>=1.11",
    "pre-commit>=3.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
